Changes in 2 .js files and userpage.html
recommendation.js had a change and delete it.
survey.js  Parses the self rating nutritional data from the form and
updates tags rating and uses the tag scores from the nutritional
recommendation form data to generate a tag string, but it is missing the
code for parsesing the nutritional recommendation data from the form and
updates tags .
userpage.html hada some css changes.